+++ tests/specs/references/exportAssignment/default/intermediateVariable/class.d.ts
@@ -1,0 +1,5 @@
+declare const chalk: ChalkInstance;
+
+export declare class ChalkInstance {}
+
+export default chalk;